requirements.txt

fastapi
uvicorn[standard]
python-multipart
python-dotenv 


torch 
numpy

openai-whisper

transformers
requests
bert-extractive-summarizer